separated camera from running game state
