+++ play.py
@@ -2,6 +2,7 @@ import sys
 
 import pygame
 
+from camera import Camera
 from square import Square
 from state import GameState
 from tiles import Ground
@@ -9,61 +10,6 @@ from world import TILE_SIZE, WORLD_WIDTH
 from world_builder import world_list
 
 FPS_LIMIT = 250
-
-
-class Camera:
-    def __init__(self, SCREEN_WIDTH, SCREEN_HEIGHT, square, tiles, ground_tiles):
-        """Inicjalizuje kamerę."""
-        self.SCREEN_WIDTH = SCREEN_WIDTH
-        self.SCREEN_HEIGHT = SCREEN_HEIGHT
-        self.square = square
-        self.tiles = tiles
-        self.zoom_level = 1
-        self.target_zoom_level = 1  # Dla płynnego zoomu
-
-        # Ustaw przesunięcie kamery na środek świata gry
-        self.camera_offset_x = -WORLD_WIDTH / 2 + SCREEN_WIDTH / 2
-        self.camera_offset_y = 0
-
-        # Znajdź najniższy rząd kafelków Ground
-        self.ground_tiles = ground_tiles
-        self.lowest_row = max(tile.y for tile in ground_tiles)
-
-    def calculate_target_offset(self):
-        """Oblicza przesunięcie kamery, aby śledzić kwadrat."""
-        half_screen_width = self.SCREEN_WIDTH / 2
-        half_screen_height = self.SCREEN_HEIGHT / 2
-        self.target_offset_x = half_screen_width - (self.square.x + self.square.size / 2) * self.zoom_level
-        self.target_offset_y = half_screen_height - (self.square.y + self.square.size / 2) * self.zoom_level
-
-    def limit_target_offset(self):
-        """Ogranicza przesunięcie kamery, aby nie wyświetlać obszarów poza światem gry."""
-        ground_tiles = [tile for tile in self.tiles if isinstance(tile, Ground)]
-        lowest_row = max(tile.y for tile in ground_tiles)
-        if self.target_offset_y < -lowest_row * self.zoom_level + self.SCREEN_HEIGHT - TILE_SIZE * self.zoom_level:
-            self.target_offset_y = -lowest_row * self.zoom_level + self.SCREEN_HEIGHT - TILE_SIZE * self.zoom_level
-        if self.target_offset_x > 0:
-            self.target_offset_x = 0
-        elif self.target_offset_x < self.SCREEN_WIDTH - WORLD_WIDTH * self.zoom_level:
-            self.target_offset_x = self.SCREEN_WIDTH - WORLD_WIDTH * self.zoom_level
-
-    def update_camera_offset(self):  # MOVE THIS TO CAMERA CLASS
-        """Aktualizuje przesunięcie kamery, interpolując je do docelowego przesunięcia kamery."""
-        self.camera_offset_x = self.target_offset_x
-        self.camera_offset_y = self.target_offset_y
-
-    def update_zoom(self):  # MOVE THIS TO CAMERA CLASS
-        """Aktualizuje poziom zoomu, interpolując go do docelowego poziomu zoomu."""
-        lerp_speed = 0.1  # Szybkość interpolacji, możesz dostosować tę wartość
-        self.zoom_level += (self.target_zoom_level - self.zoom_level) * lerp_speed
-
-    def handle_scroll_zoom(self, event):  # MOVE THIS TO CAMERA CLASS
-        """
-        Obsługuje zoom przy użyciu rolki myszy."""
-        if event.button == 4:
-            self.target_zoom_level *= 1.2
-        elif event.button == 5:
-            self.target_zoom_level /= 1.2
 
 
 class RunningGameState(GameState):
@@ -87,11 +33,6 @@ class RunningGameState(GameState):
 
         # Utwórz instancję Camera
         self.camera = Camera(SCREEN_WIDTH, SCREEN_HEIGHT, self.square, self.tiles, ground_tiles)
-
-    def update_camera(self):  # MOVE THIS TO CAMERA CLASS
-        self.camera.calculate_target_offset()
-        self.camera.limit_target_offset()
-        self.camera.update_camera_offset()
 
     def set_pause_state(self, pause_state):
         """Ustawia stan pauzy dla stanu gry."""
@@ -157,7 +98,7 @@ class RunningGameState(GameState):
         """Aktualizuje logikę gry dla bieżącego stanu gry."""
         self.square.update()  # Aktualizacja kwadratu
         self.camera.update_zoom()  # Aktualizacja zoomu
-        self.update_camera()  # Aktualizacja kamery
+        self.camera.update_camera()  # Aktualizacja kamery
         self.handle_ground_collisions()  # Sprawdź kolizje między kwadratem a wszystkimi kafelkami
 
     def draw(self, screen):
